Доработки по серверной части (#3)
* Добавил основу для сервера, поправил README.md

* небольшие правки в скрипте запуска

* gitignore

* gitignore update 2

* Модель, миграции, запрос на вставку полученных данных, index.html, views, urls

* gitignore

* изменение типа поля модели + миграции

* Сверстал html, стили, скрипт для автоподгрузки и агрегированных данных(js)

* Добавил метод на получение 100 последних записей из таблицы, использовал метод в эндпоинте "/".;

Добавил эндпоинт для получения этих данных.

* requirements.txt

* README.md, скрипт для запуска сервера

* README.md, скрипт запуска

+++ server/cpu_checker/cpu_checker_service/models.py
@@ -2,4 +2,4 @@ from django.db import models
 
 
 class CpuData(models.Model):
-    cpu_utilization = models.DecimalField(max_digits=4, decimal_places=1, verbose_name='Утилизация ЦП, %')
+    cpu_utilization = models.IntegerField(verbose_name='Утилизация ЦП, %')
